refactor: Rename mappers
