+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/data/mapper/PlaceMapper.kt
@@ -1,0 +1,18 @@
+package com.teamoffroad.feature.explore.data.mapper
+
+import com.teamoffroad.feature.explore.data.remote.response.PlaceResponseDto
+import com.teamoffroad.feature.explore.domain.model.Place
+
+fun PlaceResponseDto.toDomain(): Place {
+    return Place(
+        id = id,
+        name = name,
+        address = address,
+        shortIntroduction = shortIntroduction,
+        placeCategory = placeCategory,
+        categoryImageUrl = categoryImageUrl,
+        latitude = latitude,
+        longitude = longitude,
+        visitCount = visitCount,
+    )
+}
